backend/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.5
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.1.3
fastapi==0.104.1
future==0.18.3
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
junos-eznc==2.6.8
lxml==4.9.3
MarkupSafe==2.1.3
napalm==4.1.0
ncclient==0.6.13
netaddr==0.9.0
netmiko==4.2.0
netutils==1.6.0
ntc_templates==4.0.1
orjson==3.9.10
paramiko==3.3.1
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.5.0
pycparser==2.21
pydantic==2.5.0
pydantic-extra-types==2.1.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
pyeapi==1.0.2
PyNaCl==1.5.0
pyparsing==3.1.1
pyserial==3.5
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
scp==0.14.5
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
textfsm==1.1.3
transitions==0.9.0
ttp==0.9.5
ttp-templates==0.3.5
typing_extensions==4.8.0
ujson==5.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
yamlordereddictloader==0.4.2
